Make tab changes no longer cancel library loading
